completed exercise1-11 with test cases and variations
Completed exercise1-11 and test cases, and added exercise1-11_v1.c and
exercise1-11_v2.c to personal exercises.

NOTICE:
From here on out, unless the personal program is completely original, it will
be included in the folder with its parent program, and only a symbolic link
to the file will be included in the (/personalexercises/) folder for each
chapter.